+++ source/state_representation/src/space/joint/JointTorques.cpp
@@ -141,13 +141,26 @@ JointState JointTorques::operator+(const JointState& state) const {
   return this->JointState::operator+(state);
 }
 
+JointTorques JointTorques::operator-() const {
+  return this->JointState::operator-();
+}
+
 JointTorques& JointTorques::operator-=(const JointTorques& torques) {
   this->JointState::operator-=(torques);
   return (*this);
 }
 
+JointTorques& JointTorques::operator-=(const JointState& state) {
+  this->JointState::operator-=(state);
+  return (*this);
+}
+
 JointTorques JointTorques::operator-(const JointTorques& torques) const {
   return this->JointState::operator-(torques);
+}
+
+JointState JointTorques::operator-(const JointState& state) const {
+  return this->JointState::operator-(state);
 }
 
 std::ostream& operator<<(std::ostream& os, const JointTorques& torques) {