+++ source/state_representation/src/space/joint/JointVelocities.cpp
@@ -180,13 +180,26 @@ JointState JointVelocities::operator+(const JointState& state) const {
   return this->JointState::operator+(state);
 }
 
+JointVelocities JointVelocities::operator-() const {
+  return this->JointState::operator-();
+}
+
 JointVelocities& JointVelocities::operator-=(const JointVelocities& velocities) {
   this->JointState::operator-=(velocities);
   return (*this);
 }
 
+JointVelocities& JointVelocities::operator-=(const JointState& state) {
+  this->JointState::operator-=(state);
+  return (*this);
+}
+
 JointVelocities JointVelocities::operator-(const JointVelocities& velocities) const {
   return this->JointState::operator-(velocities);
+}
+
+JointState JointVelocities::operator-(const JointState& state) const {
+  return this->JointState::operator-(state);
 }
 
 std::ostream& operator<<(std::ostream& os, const JointVelocities& velocities) {